@@ -1,30 +1,26 @@
 {
   "name": "ts-frontend-boilerplate",
+  "private": true,
   "version": "1.0.0",
   "description": "A Boilerplate project for setting up a simple react Front End with Typescript and other things I like",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no tests installed\" && exit 1",
-    "build:dev": "webpack --config webpack.config.js --mode=development",
-    "start": "webpack-dev-server --mode development --open --hot"
+    "build": "tsc && vite build",
+    "dev": "vite",
+    "preview": "vite preview"
   },
+  "type": "module",
   "author": "Peter Lockett",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.17.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/react": "^16.9.19",
-    "@types/react-dom": "^16.9.5",
-    "awesome-typescript-loader": "^5.2.1",
-    "clean-webpack-plugin": "^3.0.0",
-    "html-webpack-plugin": "^3.2.0",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.5",
-    "webpack": "^4.41.5",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.2"
+    "@types/react": "^18.2.47",
+    "@types/react-dom": "^18.2.18",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "typescript": "^5.3.3",
+    "vite": "^5.0.11"
   }
 }
