@@ -1,8 +1,9 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import {App} from "./app";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from "./app/app";
 
-ReactDOM.render(
-    <App/>,
-    document.getElementById('app')
+ReactDOM.createRoot(document.getElementById('app')!).render(
+    <React.StrictMode>
+        <App/>
+    </React.StrictMode>
 );
